moving code to services_rules.py
