+++ api/mapping/services.py
@@ -8,7 +8,6 @@ from django.contrib import messages
 import json
 import requests
 import time
-from data.models import Concept, ConceptRelationship
 from .models import (
     ScanReport,
     ScanReportTable,
@@ -17,66 +16,6 @@ from .models import (
     DataDictionary
 )
 
-from coconnect.tools.omop_db_inspect import OMOPDetails
-
-
-
-def get_concept_from_concept_code(concept_code,
-                                  vocabulary_id,
-                                  no_source_concept=False):
-    """
-    Given a concept_code and vocabularly id, 
-    return the source_concept and concept objects
-
-    If the concept is a standard concept, 
-    source_concept will be the same object
-
-    Parameters:
-      concept_code (str) : the concept code  
-      vocabulary_id (str) : SNOMED etc.
-      no_source_concept (bool) : only return the concept
-    Returns:
-      tuple( source_concept(Concept), concept(Concept) )
-      OR
-      concept(Concept)
-    """
-    #obtain the source_concept given the code and vocab
-    source_concept = Concept.objects.get(
-        concept_code = concept_code,
-        vocabulary_id = vocabulary_id
-    )
-
-    #if the source_concept is standard
-    if source_concept.standard_concept == 'S':
-        #the concept is the same as the source_concept
-        concept = source_concept
-    else:
-        #otherwise we need to look up 
-        concept = find_standard_concept(source_concept)
-
-    if no_source_concept:
-        #only return the concept
-        return concept
-    else:
-        #return both as a tuple
-        return (source_concept,concept)
-
-
-def find_standard_concept(source_concept):
-
-    concept_relation = ConceptRelationship.objects.get(
-        concept_id_1=source_concept.concept_id,
-        relationship_id__contains='Maps to'
-    )
-
-    if concept_relation.concept_id_2 != concept_relation.concept_id_1:
-        concept = Concept.objects.get(
-            concept_id=concept_relation.concept_id_2
-        )
-        return concept
-    else:
-        #may need some warning if this ever happens?
-        return source_concept
 
 def process_scan_report_sheet_table(filename):
     """
